Prevent MSVC warning C26451: Arithmetic overflow
MSVC will report, when static analyzer warnings are enabled:
C26451: Arithmetic overflow: Using operator '+' on a 4 byte value and then casting the result to a 8 byte value. Cast the value to the wider type before calling operator '+' to avoid overflow (io.2).

As last_char_index + 1 fits normally in an int, there is no risk to overflow, but the warning is anoying.
Fix the warning by first doing the cast and then adding 1.